% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{as.data.frame.grafeo_result}
\alias{as.data.frame.grafeo_result}
\title{Convert a Grafeo result to a data frame}
\usage{
\method{as.data.frame}{grafeo_result}(x, ...)
}
\arguments{
\item{x}{A \code{grafeo_result}.}

\item{...}{Unused.}
}
\value{
A base \code{data.frame}.
}
\description{
Convert a Grafeo result to a data frame
}
