% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafeo-package.R
\docType{package}
\name{grafeoR-package}
\alias{grafeoR}
\alias{grafeoR-package}
\title{grafeoR}
\description{
R bindings for the embedded Grafeo graph database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hadimaster65555.github.io/grafeoR/}
  \item \url{https://github.com/hadimaster65555/grafeoR}
  \item \url{https://github.com/GrafeoDB/grafeo}
  \item \url{https://grafeo.dev/}
  \item Report bugs at \url{https://github.com/hadimaster65555/grafeoR/issues}
}

}
\author{
\strong{Maintainer}: Raden Muhammad Hadi \email{31847323+hadimaster65555@users.noreply.github.com}

}
\keyword{internal}
