% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafeo.R
\name{with_grafeo_db}
\alias{with_grafeo_db}
\title{Run an expression with an automatically closed database}
\usage{
with_grafeo_db(code, path = NULL, ...)
}
\arguments{
\item{code}{An expression evaluated with \code{db} available in the calling
environment.}

\item{path}{Database path passed to \code{\link[=grafeo_db]{grafeo_db()}}.}

\item{...}{Additional arguments passed to \code{\link[=grafeo_db]{grafeo_db()}}.}
}
\value{
The value of \code{code}.
}
\description{
Run an expression with an automatically closed database
}
